Remove the Containers.h file, copied couple of methods into Weld.h and changed weld to use std::vector
