+++ src/collider/GeomTree.cpp
@@ -11,7 +11,7 @@ static const unsigned int IGNORE_FLAG = 0x8000;
 GeomTree::~GeomTree()
 {
 }
-#pragma optimize("",off)
+
 GeomTree::GeomTree(const int numVerts, const int numTris, const std::vector<vector3f> &vertices, const Uint16 *indices, const unsigned int *triflags)
 : m_numVertices(numVerts)
 , m_numTris(numTris)
@@ -51,82 +51,24 @@ GeomTree::GeomTree(const int numVerts, const int numTris, const std::vector<vect
 	else if ((_i1) > (_i2)) edges[std::pair<int,int>(_i2,_i1)] = _triflag;
 
 	// eliminate duplicate vertices
-#if 1
-	const Uint32 count = numVerts;
-	nv::Array<Uint32> xrefs;
-	nv::Array<vector3f> verts;
-	verts.reserve(numVerts);
-	for (int i = 0; i < numVerts; i++)
 	{
-		verts.push_back(m_vertices[i]);
-	}
-	nv::Weld<vector3f> weld;
-	const Uint32 newCount = weld(verts, xrefs);
+		const Uint32 count = numVerts;
+		std::vector<Uint32> xrefs;
+		nv::Weld<vector3f> weld;
+		const Uint32 newCount = weld(m_vertices, xrefs);
 
-	Output("---   %d vertices welded\n", count - newCount);
+		Output("---   %d vertices welded\n", count - newCount);
 
-	// Remap faces.
-	const Uint32 faceCount = numTris;
-	std::vector<Uint16> faces(m_indices);
-	for (Uint32 f = 0; f < faceCount; f++)
-	{
-		const Uint32 idx = (f * 3);
-		faces[idx+0] = xrefs.at(faces[idx+0]);
-		faces[idx+1] = xrefs.at(faces[idx+1]);
-		faces[idx+2] = xrefs.at(faces[idx+2]);
-	}
-#if 0
-	m_indices = faces;
-#else
-	for (int i=0; i<numVerts; i++) 
-	{
-		const vector3f &v1 = m_vertices[i];
-		for (int j = i + 1; j<numVerts; j++)
+		// Remap faces.
+		const Uint32 faceCount = numTris;
+		for (Uint32 f = 0; f < faceCount; f++)
 		{
-			const vector3f &v2 = m_vertices[j];
-			if (v2.ExactlyEqual(v1))
-			{
-				for (int k = 0; k<numIndices; k++)
-				{
-					if ((indices[k] == j) && (triflags[k / 3] < IGNORE_FLAG)) {
-						m_indices[k] = i;
-					}
-				}
-			}
+			const Uint32 idx = (f * 3);
+			m_indices[idx+0] = xrefs.at(m_indices[idx+0]);
+			m_indices[idx+1] = xrefs.at(m_indices[idx+1]);
+			m_indices[idx+2] = xrefs.at(m_indices[idx+2]);
 		}
 	}
-
-	assert(faces.size() == m_indices.size());
-	Output("Indices\n");
-	for (size_t i = 0; i < faces.size(); i++)
-	{
-		const Uint32 fi = faces[i];
-		const Uint32 ii = m_indices[i];
-		Output("F(%4d):v(%9.4f,%9.4f,%9.4f) & I(%4d):v(%9.4f,%9.4f,%9.4f)\n", 
-			faces[i], m_vertices[fi].x, m_vertices[fi].y, m_vertices[fi].z,
-			m_indices[i], m_vertices[ii].x, m_vertices[ii].y, m_vertices[ii].z);
-	}
-	Output("End-of Indices\n");
-#endif
-#else
-	for (int i=0; i<numVerts; i++) 
-	{
-		const vector3f &v1 = m_vertices[i];
-		for (int j=i+1; j<numVerts; j++) 
-		{
-			const vector3f &v2 = m_vertices[j];
-			if (v2.ExactlyEqual(v1)) 
-			{
-				for (int k=0; k<numIndices; k++) 
-				{
-					if ((indices[k] == j) && (triflags[k / 3] < IGNORE_FLAG)) {
-						m_indices[k] = i;
-					}
-				}
-			}
-		}
-	}
-#endif
 
 	// Get radius, m_aabb, and merge duplicate edges
 	m_radius = 0;
