+++ src/collider/Weld.h
@@ -4,7 +4,6 @@
 #define NV_MESH_WELD_H
 
 #include "libs.h"
-#include "Containers.h"
 
 // Weld function to remove array duplicates in linear time using hashing.
 
@@ -22,6 +21,63 @@ template <class T> struct Equal
 /// Null index. @@ Move this somewhere else... This could have collisions with other definitions!
 #define NIL Uint32(~0)
 
+template <typename Key> struct hash 
+{
+	inline Uint32 sdbm_hash(const void * data_in, Uint32 size, Uint32 h = 5381)
+	{
+		const Uint8 * data = (const Uint8 *) data_in;
+		Uint32 i = 0;
+		while (i < size) {
+			h = (h << 16) + (h << 6) - h + (Uint32) data[i++];
+		}
+		return h;
+	}
+		
+	Uint32 operator()(const Key & k) {
+		return sdbm_hash(&k, sizeof(Key));
+	}
+};
+template <> struct hash<int>
+{
+	Uint32 operator()(int x) const { return x; }
+};
+template <> struct hash<Uint32>
+{
+	Uint32 operator()(Uint32 x) const { return x; }
+};
+
+/** Return the next power of two. 
+* @see http://graphics.stanford.edu/~seander/bithacks.html
+* @warning Behaviour for 0 is undefined.
+* @note isPowerOfTwo(x) == true -> nextPowerOfTwo(x) == x
+* @note nextPowerOfTwo(x) = 2 << log2(x-1)
+*/
+inline Uint32 nextPowerOfTwo( Uint32 x )
+{
+	assert( x != 0 );
+#if 1	// On modern CPUs this is supposed to be as fast as using the bsr instruction.
+	x--;
+	x |= x >> 1;
+	x |= x >> 2;
+	x |= x >> 4;
+	x |= x >> 8;
+	x |= x >> 16;
+	return x+1;	
+#else
+	Uint32 p = 1;
+	while( x > p ) {
+		p += p;
+	}
+	return p;
+#endif
+}
+
+/// Return true if @a n is a power of two.
+inline bool isPowerOfTwo( Uint32 n )
+{
+	return (n & (n-1)) == 0;
+}
+
 
 /// Generic welding routine. This function welds the elements of the array p
 /// and returns the cross references in the xrefs array. To compare the elements
@@ -32,7 +88,7 @@ template <class T, class H=hash<T>, class E=Equal<T> >
 struct Weld
 {
 	// xrefs maps old elements to new elements
-	Uint32 operator()(Array<T> & p, Array<Uint32> & xrefs)
+	Uint32 operator()(std::vector<T> & p, std::vector<Uint32> & xrefs)
 	{
 		const Uint32 N = p.size();							// # of input vertices.
 		Uint32 outputCount = 0;								// # of output vertices
@@ -41,14 +97,15 @@ struct Weld
 		Uint32 * next = hashTable + hashSize;					// use bottom part as linked list
 
 		xrefs.resize(N);
-		memset( hashTable, NIL, hashSize*sizeof(Uint32) );	// init hash table (NIL = 0xFFFFFFFF so memset works)
+		memset( hashTable, NIL, (hashSize + N)*sizeof(Uint32) );	// init hash table (NIL = 0xFFFFFFFF so memset works)
 
 		H hash;
 		E equal;
 		for (Uint32 i = 0; i < N; i++)
 		{
 			const T & e = p[i];
-			Uint32 hashValue = hash(e) & (hashSize-1);
+			const Uint32 hashValue = hash(e) & (hashSize-1);
+			//const Uint32 hashValue = CodeSupHash(e) & (hashSize-1);
 			Uint32 offset = hashTable[hashValue];
 
 			// traverse linked list
@@ -85,33 +142,6 @@ struct Weld
 		return outputCount;
 	}
 };
-
-
-/// Reorder the given array accoding to the indices given in xrefs.
-template <class T>
-void reorderArray(Array<T> & array, const Array<Uint32> & xrefs)
-{
-	const Uint32 count = xrefs.count();
-	Array<T> new_array(count);
-
-	for(Uint32 i = 0; i < count; i++) {
-		new_array[i] = array[xrefs[i]];
-	}
-
-	swap(array, new_array);
-}
-
-/// Reverse the given array so that new indices point to old indices.
-inline void reverseXRefs(Array<Uint32> & xrefs, Uint32 count)
-{
-	Array<Uint32> new_xrefs(count);
-	
-	for(Uint32 i = 0; i < xrefs.count(); i++) {
-		new_xrefs[xrefs[i]] = i;
-	}
-	
-	swap(xrefs, new_xrefs);
-}
 
 } // nv namespace
