+++ win32/vc2013Express/collider/collider.vcxproj
@@ -135,7 +135,6 @@
     <ClInclude Include="..\..\..\src\collider\collider.h" />
     <ClInclude Include="..\..\..\src\collider\CollisionContact.h" />
     <ClInclude Include="..\..\..\src\collider\CollisionSpace.h" />
-    <ClInclude Include="..\..\..\src\collider\Containers.h" />
     <ClInclude Include="..\..\..\src\collider\Geom.h" />
     <ClInclude Include="..\..\..\src\collider\GeomTree.h" />
     <ClInclude Include="..\..\..\src\collider\Weld.h" />
